test/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-CORS
python-dotenv